backend/requirements.txt
flask==2.4.0
flask-cors==3.1.13
pandas==2.1.1
numpy==1.26.4
scikit-learn==1.3.2
matplotlib==3.8.1
joblib==1.3.2
requests==2.32.1
apscheduler==3.12.1
python-dotenv==1.0.1
pymongo==4.5.1
pytz==2025.7
gunicorn==22.3.2
